refactor: remove legacy data-driven ability execution path
- Deleted AbilityExecutor.cs (ProcessActive, TryStart, CanBufferCombo)
- Removed ProcessAttack() from Simulation.SimulateTick
- Removed StartAttackFromSlot() warp/combo logic (now ServerAbility responsibility)
- Moved GetCooldown/SetCooldown helpers to ServerSimulation
- All abilities now MUST have AbilityTypeId > 0 (ServerAbility dispatch)
